+++ CHANGELOG.md
@@ -22,6 +22,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#2044](https://github.com/open-telemetry/opentelemetry-python/pull/2044))
 - `opentelemetry-sdk` Add support for `OTEL_ATTRIBUTE_VALUE_LENGTH_LIMIT` env var
   ([#2056](https://github.com/open-telemetry/opentelemetry-python/pull/2056))
+- `opentelemetry-sdk` Treat limit even vars set to empty values as unset/unlimited.
+  ([#2054](https://github.com/open-telemetry/opentelemetry-python/pull/2054))
 - `opentelemetry-api` Attribute keys must be non-empty strings.
   ([#2057](https://github.com/open-telemetry/opentelemetry-python/pull/2057))